// WYNKWeb/src/app/Models/InvestigationBillMaster.model.ts
export class InvestigationBillMaster {

  CMPID: number;
  RegistrationTranID: number;
  UIN: string;
  BillNo: string;
  TransactionId: number;
  GrossProductValue: number;
  TotalDiscountValue: number;
  TotalTaxValue: number;
  TotalCGSTTaxValue: number;
  TotalSGSTTaxValue: number;
  TotalBillValue: number;
  CESSAmount: number;
  AdditionalCESSAmount: number;
  TransactionID: number;
  CreatedBy: number;
  TaxID: number;
}